container/requirements.txt
Flask
numpy
pillow
cassandra-driver